+++ app/ai-powered-email-analyzer/page.tsx
@@ -4,10 +4,6 @@ import { Brain, BarChart, Target, TrendingUp, ArrowRight, CheckCircle } from 'lu
 import Navigation from '../components/Navigation';
 import Footer from '../components/Footer';
 
-import React from 'react';
-import { Brain, BarChart, Target, TrendingUp, ArrowRight, CheckCircle } from 'lucide-react';
-import Navigation from '../components/Navigation';
-import Footer from '../components/Footer';
 
 export const metadata = {
   title: 'AI-Powered Email Analyzer | Zion Tech Group',
