+++ app/it-services/cybersecurity-audit/page.tsx
@@ -3,9 +3,6 @@ import React from 'react';
 import Navigation from '../../components/Navigation';
 import Footer from '../../components/Footer';
 
-import React from 'react';
-import Navigation from '../../components/Navigation';
-import Footer from '../../components/Footer';
 
 export const metadata = {
   title: 'Cybersecurity Audit | Zion Tech Group',
