+++ app/medical-records-manager/page.tsx
@@ -4,8 +4,6 @@ import Navigation from '../components/Navigation';
 import Footer from '../components/Footer';
 
 
-import Navigation from '../components/Navigation';
-import Footer from '../components/Footer';
 
 export const metadata = {
   title: 'Medical Records Manager | Zion Tech Group',
