+++ app/online-learning-platform/page.tsx
@@ -4,8 +4,6 @@ import Navigation from '../components/Navigation';
 import Footer from '../components/Footer';
 
 
-import Navigation from '../components/Navigation';
-import Footer from '../components/Footer';
 
 export const metadata = {
   title: 'Online Learning Platform | Zion Tech Group',
