+++ app/supply-chain-optimizer/page.tsx
@@ -4,8 +4,6 @@ import Navigation from '../components/Navigation';
 import Footer from '../components/Footer';
 
 
-import Navigation from '../components/Navigation';
-import Footer from '../components/Footer';
 
 export const metadata = {
   title: 'Supply Chain Optimizer | Zion Tech Group',
